Refactor reactUtils.ts by extracting focused utility modules
Split the monolithic reactUtils.ts (518 lines, 20+ utilities) into focused modules following Single Responsibility Principle:

Created modules:
- focusUtils.ts: Focus management, tab trapping, and focus event detection
- keyboardUtils.ts: Keyboard event handling and key combinations
- mediaQueryUtils.ts: Media queries and responsive design hooks
- scrollUtils.ts: Scroll position monitoring

Updated reactUtils.ts:
- Now serves as a barrel export re-exporting from new modules
- Retains general-purpose utilities (DOM events, timeouts, SSR detection)
- Maintains backward compatibility - all existing imports continue to work

Benefits:
- Better code organization by concern
- Easier to find and maintain related utilities
- Comprehensive JSDoc comments added to all functions
- Each module is now testable in isolation
- Reduced file size improves readability

No breaking changes - all existing exports preserved through re-exports.

Related to CORE-1452 and CORE-1454
